Refactor test utilities into a module
