+++ tests/lib/rules/enforce-import-ordering.js
@@ -1,8 +1,6 @@
-const RuleTester = require('eslint').RuleTester;
+const {compileTests, getRuleTester} = require('../utils.js');
 
 const rule = require('../../../lib/rules/enforce-import-ordering');
-
-const ruleTester = new RuleTester({parserOptions: {ecmaVersion: 2022, sourceType: 'module'}});
 
 const testSets = {
   ordering: () => {
@@ -218,31 +216,4 @@ const testSets = {
   }
 };
 
-const attachOptions = (tests, options) =>
-  options ? tests.map(test => ({...test, options: [test.options || options]})) : tests
-
-const compileTests = () => {
-
-  // Compile the expected tests structure from the internal hierarchical test
-  // sets structure. The latter is an object whose every value is a function
-  // returning a test set containing `valid` and `invalid` test cases, and:
-  //
-  // - options - common options applied to each test case, unless the test
-  //   case has its own `options`
-
-  return Object.values(testSets).reduce(
-    (acc, makeTestSet) => {
-      const testSet = makeTestSet();
-      const valid = attachOptions(testSet.valid, testSet.options);
-      const invalid = attachOptions(testSet.invalid, testSet.options);
-
-      return {
-        valid: [...acc.valid, ...valid],
-        invalid: [...acc.invalid, ...invalid]
-      };
-    },
-    {valid: [], invalid: []}
-  );
-}
-
-ruleTester.run('enforce-import-ordering', rule, compileTests());
+getRuleTester().run('enforce-import-ordering', rule, compileTests(testSets));
